[BWC] updates to BWC test scripts (#1190)
Allows the calling script with file location or url to enable
the CI/CD pipeline to pass URLs.

Some minor clean ups related to linter and pass by CSV for params.

Finally, update to add releases parameter which enables us to test
version mismatch if/when we allow the out of the box experience of
OpenSearch Dashboards to work with previous versions of OpenSearch.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>