[CI] allow for manual triggered jobs to accept inputs
Utilizing the GitHub `workflow_dispatch`, we can trigger
a job with a specific input.

Here we are allowing for tests to be ran pulling down from:
https://github.com/opensearch-project/opensearch-dashboards-functional-test

This workflow uses the target branch of the PR to pull down
from the FTRepo and run the tests from that specific branch.

For example, PRs against `main` will pull down `main` from
the FTRepo.

The problem occurs when new functionality is opened or a bug is
fixed and, per industry standard, we want to see tests added
to ensure functionality, stability, and raise the bar.

But the cypress tests PR into the FTRepo depends on the new code
to be merged for the CI within the FTRepo to work. So we have a
stalemate that usually slows down PR review time.

Here, a manual run can be triggered using the branch provided by
the maintainer who ran the job.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>